Engram-AiMemory/pyproject.toml
[project]
name = "ai-memory-system"
version = "1.1.0"
description = "3-Tier AI Memory System with Weaviate, Redis, and MCP"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "AI Memory Team" }]
keywords = ["weaviate", "memory", "mcp", "ai", "vector-database"]

dependencies = [
    "weaviate-client>=4.9.0",
    "redis>=5.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "mcp>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "tenacity>=8.0.0",
    "orjson>=3.10.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "slowapi>=0.1.9",
    "sentence-transformers>=3.0.0",
    "apscheduler>=3.10.0",
    "crawl4ai>=0.7.8",
    "PyMuPDF>=1.24.0",
    "pytesseract>=0.3.13",
    "pdf2image>=1.17.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.2",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.0.0",
    "black>=24.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["packages/core/src/memory_system"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Enable gradually
disallow_incomplete_defs = false  # Enable gradually
check_untyped_defs = true
ignore_missing_imports = true
# Phase 2: Remove these suppressions as code is typed
disable_error_code = ["no-any-return", "type-arg", "assignment", "var-annotated", "misc"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["packages/core/tests"]
addopts = "--cov=memory_system --cov-report=term-missing --cov-report=html --cov-report=xml --cov-fail-under=79.8"

[tool.coverage.run]
omit = [
    "*/src/memory_system/mcp/*",
    "*/src/memory_system/investigation/crawler.py",
    "*/src/memory_system/investigation/crawler_service.py",
    "*/src/memory_system/investigation/ingestor.py",
    "*/src/memory_system/investigation/workers.py",
    "*/src/memory_system/investigation/workers_service.py",
    "*/src/memory_system/investigation/schemas.py",
    "*/src/memory_system/compat.py",
    "*/src/memory_system/ollama_client.py",
    "*/src/memory_system/ai_provider.py",
]
